--- cmd/pulumi-resource-dnsimple/schema.json
@@ -1,1 +1,0 @@
-{"name":"dnsimple","version":"v1.2.0+dirty","description":"A Pulumi package for creating and managing dnsimple cloud resources.\n\n\u003e This provider is a derived work of the [Terraform Provider](https://github.com/terraform-providers/terraform-provider-dnsimple)\n\u003e distributed under [MPL 2.0](https://www.mozilla.org/en-US/MPL/2.0/). If you encounter a bug or missing feature,\n\u003e first check the [`pulumi/pulumi-dnsimple` repo](https://github.com/pulumi/pulumi-dnsimple/issues); however, if that doesn't turn up anything,\n\u003e please consult the source [`terraform-providers/terraform-provider-dnsimple` repo](https://github.com/terraform-providers/terraform-provider-dnsimple/issues).","keywords":["pulumi","dnsimple"],"homepage":"https://pulumi.io","license":"Apache-2.0","repository":"https://github.com/pulumi/pulumi-dnsimple","meta":{"moduleFormat":"(.*)(?:/[^/]*)"},"config":{"variables":{"account":{"type":"string","description":"The account for API operations.\n","default":"","defaultInfo":{"environment":["DNSIMPLE_ACCOUNT"]}},"token":{"type":"string","description":"The API v2 token for API operations.\n","default":"","defaultInfo":{"environment":["DNSIMPLE_TOKEN"]}}},"defaults":["account","token"]},"provider":{"description":"The provider type for the dnsimple package. By default, resources use package-wide configuration\nsettings, however an explicit `Provider` instance may be created and passed during resource\nconstruction to achieve fine-grained programmatic control over provider settings. See the\n[documentation](https://www.pulumi.com/docs/reference/programming-model/#providers) for more information.\n\n\u003e This content is derived from https://github.com/terraform-providers/terraform-provider-dnsimple/blob/master/website/docs/index.html.markdown.\n","inputProperties":{"account":{"type":"string","description":"The account for API operations.\n","default":"","defaultInfo":{"environment":["DNSIMPLE_ACCOUNT"]}},"token":{"type":"string","description":"The API v2 token for API operations.\n","default":"","defaultInfo":{"environment":["DNSIMPLE_TOKEN"]}}}},"resources":{"dnsimple:index/record:Record":{"description":"Provides a DNSimple record resource.\n\n\u003e This content is derived from https://github.com/terraform-providers/terraform-provider-dnsimple/blob/master/website/docs/r/record.html.markdown.\n","properties":{"domain":{"type":"string","description":"The domain to add the record to\n"},"domainId":{"type":"string","description":"The domain ID of the record\n"},"hostname":{"type":"string","description":"The FQDN of the record\n"},"name":{"type":"string","description":"The name of the record\n"},"priority":{"type":"number","description":"The priority of the record - only useful for some record types\n"},"ttl":{"type":"number","description":"The TTL of the record\n"},"type":{"type":"string","$ref":"#/types/dnsimple:index:RecordType","description":"The type of the record\n"},"value":{"type":"string","description":"The value of the record\n"}},"required":["domain","domainId","hostname","name","priority","type","value"],"inputProperties":{"domain":{"type":"string","description":"The domain to add the record to\n"},"name":{"type":"string","description":"The name of the record\n"},"priority":{"type":"number","description":"The priority of the record - only useful for some record types\n"},"ttl":{"type":"number","description":"The TTL of the record\n"},"type":{"type":"string","$ref":"#/types/dnsimple:index:RecordType","description":"The type of the record\n"},"value":{"type":"string","description":"The value of the record\n"}},"requiredInputs":["domain","name","type","value"],"stateInputs":{"description":"Input properties used for looking up and filtering Record resources.\n","properties":{"domain":{"type":"string","description":"The domain to add the record to\n"},"domainId":{"type":"string","description":"The domain ID of the record\n"},"hostname":{"type":"string","description":"The FQDN of the record\n"},"name":{"type":"string","description":"The name of the record\n"},"priority":{"type":"number","description":"The priority of the record - only useful for some record types\n"},"ttl":{"type":"number","description":"The TTL of the record\n"},"type":{"type":"string","$ref":"#/types/dnsimple:index:RecordType","description":"The type of the record\n"},"value":{"type":"string","description":"The value of the record\n"}},"type":"object"}}},"language":{"csharp":{"namespaces":{"dnsimple":"DNSimple"},"packageReferences":{"Pulumi":"1.9.1-preview","System.Collections.Immutable":"1.6.0"}},"nodejs":{"dependencies":{"@pulumi/pulumi":"^1.0.0"},"devDependencies":{"@types/mime":"^2.0.0","@types/node":"^8.0.25"},"packageDescription":"A Pulumi package for creating and managing dnsimple cloud resources.","packageName":"","typescriptVersion":""},"python":{"requires":{"pulumi":"\u003e=1.0.0,\u003c2.0.0"}}}}
