+++ provider/cmd/pulumi-resource-dnsimple/.gitignore
@@ -1,0 +1,2 @@
+schema.json
+schema.go
