+++ provider/cmd/pulumi-resource-dnsimple/Pulumi.yaml
