+++ provider/cmd/pulumi-resource-dnsimple/generate.go
