+++ provider/cmd/pulumi-tfgen-dnsimple/main.go
@@ -17,8 +17,8 @@ package main
 import (
 	"github.com/pulumi/pulumi-terraform-bridge/pkg/tfgen"
 
-	dnsimple "github.com/pulumi/pulumi-dnsimple"
-	"github.com/pulumi/pulumi-dnsimple/pkg/version"
+	dnsimple "github.com/pulumi/pulumi-dnsimple/provider"
+	"github.com/pulumi/pulumi-dnsimple/provider/pkg/version"
 )
 
 func main() {
