+++ provider/go.mod
@@ -1,12 +1,11 @@
-module github.com/pulumi/pulumi-dnsimple
+module github.com/pulumi/pulumi-dnsimple/provider
 
 go 1.13
 
 require (
-	github.com/hashicorp/terraform-plugin-sdk v1.6.0
-	github.com/pkg/errors v0.9.1
-	github.com/pulumi/pulumi v1.12.2-0.20200313044354-8111d33438b9
-	github.com/pulumi/pulumi-terraform-bridge v1.8.2
+	github.com/hashicorp/terraform-plugin-sdk v1.7.0
+	github.com/pulumi/pulumi-terraform-bridge v1.8.4
+	github.com/pulumi/pulumi/sdk v0.0.0-20200325225746-80f1989600a3
 	github.com/terraform-providers/terraform-provider-dnsimple v0.3.0
 )
