+++ scripts/get-version
@@ -3,34 +3,51 @@ set -o nounset -o errexit -o pipefail
 SCRIPT_DIR="$( cd "$( dirname "$0" )" && pwd )"
 COMMITISH=${1:-HEAD}
 DIRTY_TAG=""
-
+EXACT=0
 # Figure out if the worktree is dirty, we run update-index first
 # as we've seen cases in Travis where not doing so causes git to
 # treat the worktree as dirty when it is not.
 git update-index -q --refresh
-if ! git diff-files --quiet; then
+if ! git diff-files --quiet -- . ':!**/go.mod' ':!**/go.sum'; then
     DIRTY_TAG="dirty"
 fi
 
 # If we have an exact tag, just use it.
 if git describe --tags --exact-match "${COMMITISH}" >/dev/null 2>&1; then
-    echo -n "$(git describe --tags --exact-match "${COMMITISH}")"
+    EXACT=1
+    TAG=$(git describe --tags --exact-match "${COMMITISH}")
+# Otherwise, increment the minor version version (if the package is 1.X or later) or the
+# patch version (if the package is pre 1.0), add the -alpha tag and some
+# commit metadata. If there's no existing tag, pretend a v0.0.0 was
+# there so we'll produce v0.0.1-dev builds.
+elif git describe --tags --abbrev=0 "${COMMITISH}" > /dev/null 2>&1; then
+    TAG=$(git describe --tags --abbrev=0 "${COMMITISH}")
+else
+    TAG="v0.0.0"
+fi
+
+# Check to see if the latest tag is for a submodule and transform to semver tolerant
+# e.g: submodule/submodule/.../v.X.X.X
+# sdk/v1.1.1 -> v1.1.1
+IFS='/'
+read -a tagsplit <<< "$TAG"
+
+if echo "${#tagsplit[@]}" > /dev/null 2>&1 -gt 1;
+then
+    TAG=${tagsplit[${#tagsplit[@]}-1]}
+fi
+
+# if we found an exact match, return it at this point
+# after we have removed and submodule prefixes.
+if [ $EXACT -eq 1 ]
+then
+    echo -n "$TAG"
     if [ ! -z "${DIRTY_TAG}" ]; then
         echo -n "+${DIRTY_TAG}"
     fi
 
     echo ""
     exit 0
-fi
-
-# Otherwise, increment the minor version version (if the package is 1.X or later) or the
-# patch version (if the package is pre 1.0), add the -alpha tag and some
-# commit metadata. If there's no existing tag, pretend a v0.0.0 was
-# there so we'll produce v0.0.1-dev builds.
-if git describe --tags --abbrev=0 "${COMMITISH}" > /dev/null 2>&1; then
-    TAG=$(git describe --tags --abbrev=0 "${COMMITISH}")
-else
-    TAG="v0.0.0"
 fi
 
 # Strip off any pre-release tag we might have (e.g. from doing a -rc build)
