+++ scripts/go.mod
@@ -1,0 +1,3 @@
+module github.com/pulumi/pulumi-dnsimple/scripts
+
+go 1.13
