+++ sdk/go.mod
@@ -1,0 +1,8 @@
+module github.com/pulumi/pulumi-dnsimple/sdk
+
+go 1.13
+
+require (
+	github.com/pkg/errors v0.9.1
+	github.com/pulumi/pulumi/sdk v1.14.0
+)
